Hebrew translation
I've also added a Crwodin project for that, would you like to further integrate with it and have full permissions or you would rather me deleting it? 

It's auto updated at the moment so when new translations are presented they will be added to the platfrom automatically while all the translators should get notifications about it (if this is what they've chose).
Adding a new translation currently requires some manual work but with the right configuration it can be automated.