Add action to set the timestamp for items in a meal category
Add a new action menu item to let the user manually set the timestamp (hour & minute only) of all items in a meal category. This supports all hour cycles and thus different locales. The option does not exist if the user does not have "Show Timestamps" turned on and it is disabled when there are no items in the meal category.