Use F7 method to force open the time picker on calendar open
Instead of letting the user open the time picker from the calendar we call the method to open the time picker when the calendar opens and close the calendar when the time picker closes. That way we can hide the calendar (assuming the device isn't very slow and shows the calendar for a frame).